【MindStudio ******】【系统调优】【profiling】【Ascend Insight】timeline页面下的静态统计信息区时间字段单位为ns，与设计规格us不一致
# message auto-generated for no-merge-commit merge:
merge dev1230_b1 into master

【MindStudio ******】【系统调优】【profiling】【Ascend Insight】timeline页面下的静态统计信息区时间字段单位为ns，与设计规格us不一致

Created-by: insightDeveloper

Match-id-dc6b781c0ef5ed5494c0d51ea72345c5e8c5fedb

+++ server/src/modules/timeline/protocol/TimelineProtocolResponse.h
@@ -252,12 +252,12 @@ struct UnitCounterResponse : public Response {
 struct SystemViewDetail {
     std::string name;
     double time;
-    uint64_t totalTime;
+    double totalTime;
     uint64_t numberCalls;
     double avg;
     double stdDev;
-    uint64_t min;
-    uint64_t max;
+    double min;
+    double max;
 };
 
 struct SystemViewBody {
